add condition

+++ database/migrations/2025_03_03_054816_create_client_details_table.php
@@ -31,7 +31,10 @@ return new class extends Migration
     public function down(): void
     {
         if (Schema::hasTable('client_details')) {
-            Schema::dropIfExists('client_details');
+            $count = DB::table('client_details')->count();
+            if ($count > 0) {
+                Schema::dropIfExists('client_details');
+            }   
         }
     }
 };
